Unzip and install from web
This branch changes the behavior to pull the zip file from the Zabbix
website and install the requested version. Unzipped file names will also
be have 32 or 64 added to the end to be compatible with the previous
recipe version. This checkin will also upgrade the agent when the
version is incremented.